+++ src/main/java/com/example/pdp_project/repo/UserRepository.java
@@ -2,11 +2,8 @@ package com.example.pdp_project.repo;
 
 import com.example.pdp_project.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.security.core.userdetails.UserDetails;
 
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     User findByEmail(String email);
 }
